optmised code in nationality get and post controllers
