+++ src/main/steps/applicant2/nationality/get.test.ts
@@ -1,62 +1,21 @@
-import { mockRequest } from '../../../../test/unit/utils/mockRequest';
-import { mockResponse } from '../../../../test/unit/utils/mockResponse';
+import { FieldPrefix } from '../../../app/case/case';
+import NationalityGetControllerBase from '../../../app/nationality/NationalityGetController';
 
 import { generateContent } from './content';
 import NationalityGetController from './get';
 
-describe('applicant2 > NationalityGetController', () => {
+describe('applicant2 > nationality > get', () => {
   let controller;
-  let req;
-  let res;
 
   beforeEach(() => {
     controller = new NationalityGetController(__dirname + '../../common/template', generateContent);
-    req = mockRequest();
-    res = mockResponse();
   });
 
-  describe('when there is no remove param in req.query', () => {
-    test('should not redirect', async () => {
-      await controller.get(req, res);
-      expect(res.redirect).not.toHaveBeenCalled();
-    });
+  test('should extend NationalityGetControllerBase', async () => {
+    expect(controller).toBeInstanceOf(NationalityGetControllerBase);
   });
 
-  describe('when there is no applicant2AdditionalNationalities in userCase', () => {
-    test('should not redirect', async () => {
-      req.query.remove = 'MOCK_COUNTRY';
-      await controller.get(req, res);
-      expect(res.redirect).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('when there is remove param in req.query and applicant2AdditionalNationalities in userCase', () => {
-    test('should redirect to same url', async () => {
-      req = mockRequest({
-        query: { remove: 'MOCK_COUNTRY' },
-        session: { userCase: { applicant2AdditionalNationalities: ['MOCK_COUNTRY', 'MOCK_COUNTRY2'] } },
-      });
-      req.url = '/request?remove=MOCK_COUNTRY';
-      await controller.get(req, res);
-      expect(req.session.userCase.applicant2AdditionalNationalities).toEqual(['MOCK_COUNTRY2']);
-      expect(res.redirect).toHaveBeenCalledWith('/request');
-    });
-  });
-
-  describe('when there is an error in saving session', () => {
-    test('should throw an error', async () => {
-      req = mockRequest({
-        session: {
-          user: { email: 'test@example.com' },
-          save: jest.fn(done => done('MOCK_ERROR')),
-        },
-      });
-      try {
-        await controller.get(req, res);
-      } catch (err) {
-        //eslint-disable-next-line jest/no-conditional-expect
-        expect(err).toBe('MOCK_ERROR');
-      }
-    });
+  test('should call super constructor with correct params', async () => {
+    expect(controller.fieldPrefix).toBe(FieldPrefix.APPLICANT2);
   });
 });
