+++ src/main/steps/applicant2/nationality/post.ts
@@ -1,42 +1,12 @@
 import autobind from 'autobind-decorator';
-import { Response } from 'express';
 
-import { AppRequest } from '../../../app/controller/AppRequest';
-import { AnyObject, PostController } from '../../../app/controller/PostController';
-import { Form } from '../../../app/form/Form';
-import { getNextStepUrl } from '../../../steps';
+import { FieldPrefix } from '../../../app/case/case';
+import { FormFields, FormFieldsFn } from '../../../app/form/Form';
+import NationalityPostControllerBase from '../../../app/nationality/NationalityPostController';
 
 @autobind
-export default class NationalityPostController extends PostController<AnyObject> {
-  public async post(req: AppRequest<AnyObject>, res: Response): Promise<void> {
-    const fields = typeof this.fields === 'function' ? this.fields(req.session.userCase) : this.fields;
-    const form = new Form(fields);
-    const { saveAndSignOut, saveBeforeSessionTimeout, _csrf, ...formData } = form.getParsedBody(req.body);
-
-    req.session.errors = form.getErrors(formData);
-
-    Object.assign(req.session.userCase, formData);
-
-    if (req.session.errors.length === 0) {
-      if (formData.addButton) {
-        if (!req.session.userCase.applicant2AdditionalNationalities) {
-          req.session.userCase.applicant2AdditionalNationalities = [];
-        }
-        if (formData.addAnotherNationality) {
-          req.session.userCase.applicant2AdditionalNationalities.push(formData.addAnotherNationality);
-          req.session.userCase.addAnotherNationality = '';
-        }
-      }
-    }
-
-    const nextUrl =
-      req.session.errors.length > 0 || formData.addButton ? req.url : getNextStepUrl(req, req.session.userCase);
-
-    req.session.save(err => {
-      if (err) {
-        throw err;
-      }
-      res.redirect(nextUrl);
-    });
+export default class NationalityPostController extends NationalityPostControllerBase {
+  constructor(protected readonly fields: FormFields | FormFieldsFn) {
+    super(fields, FieldPrefix.APPLICANT2);
   }
 }
